+++ src/collective/eeafaceted/collectionwidget/browser/views.py
@@ -1,11 +1,20 @@
 # -*- coding: utf-8 -*-
+
 from collective.eeafaceted.collectionwidget.content.dashboardcollection import IDashboardCollection
 from collective.eeafaceted.collectionwidget.interfaces import NoCollectionWidgetDefinedException
 from collective.eeafaceted.collectionwidget.utils import getCollectionLinkCriterion
 from eea.facetednavigation.browser.app.view import FacetedContainerView
 from eea.facetednavigation.subtypes.interfaces import IFacetedNavigable
 from plone import api
+from plone.app.contentlisting.interfaces import IContentListing
+from plone.app.querystring import queryparser
+from plone.app.querystring.interfaces import IParsedQueryIndexModifier
+from plone.app.querystring.querybuilder import logger
+from plone.app.querystring.querybuilder import QueryBuilder as OriginalQueryBuilder
+from plone.batching import Batch
+from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView
+from zope.component import getUtilitiesFor
 
 
 class RenderCategoryView(BrowserView):
@@ -88,3 +97,69 @@ class FacetedDashboardView(FacetedContainerView):
                         self.request.RESPONSE.redirect(container.absolute_url())
                         return ''
         return self.index()
+
+
+class QueryBuilder(OriginalQueryBuilder):
+    """ """
+
+    def _makequery(self, query=None, batch=False, b_start=0, b_size=30,
+                   sort_on=None, sort_order=None, limit=0, brains=False,
+                   custom_query=None):
+        """Overrided to avoid added "path" index."""
+        parsedquery = queryparser.parseFormquery(
+            self.context, query, sort_on, sort_order)
+
+        index_modifiers = getUtilitiesFor(IParsedQueryIndexModifier)
+        for name, modifier in index_modifiers:
+            if name in parsedquery:
+                new_name, query = modifier(parsedquery[name])
+                parsedquery[name] = query
+                # if a new index name has been returned, we need to replace
+                # the native ones
+                if name != new_name:
+                    del parsedquery[name]
+                    parsedquery[new_name] = query
+
+        # Check for valid indexes
+        catalog = getToolByName(self.context, 'portal_catalog')
+        valid_indexes = [index for index in parsedquery
+                         if index in catalog.indexes()]
+
+        # We'll ignore any invalid index, but will return an empty set if none
+        # of the indexes are valid.
+        if not valid_indexes:
+            logger.warning(
+                "Using empty query because there are no valid indexes used.")
+            parsedquery = {}
+
+        if not parsedquery:
+            if brains:
+                return []
+            else:
+                return IContentListing([])
+
+        if batch:
+            parsedquery['b_start'] = b_start
+            parsedquery['b_size'] = b_size
+        elif limit:
+            parsedquery['sort_limit'] = limit
+
+        # Begin changes, comment "path" aribitrary added
+        # if 'path' not in parsedquery:
+        #     parsedquery['path'] = {'query': ''}
+        # End changes, comment "path" aribitrary added
+
+        if isinstance(custom_query, dict):
+            # Update the parsed query with extra query dictionary. This may
+            # override parsed query options.
+            parsedquery.update(custom_query)
+        results = catalog(**parsedquery)
+        if getattr(results, 'actual_result_count', False) and limit\
+                and results.actual_result_count > limit:
+            results.actual_result_count = limit
+
+        if not brains:
+            results = IContentListing(results)
+        if batch:
+            results = Batch(results, b_size, start=b_start)
+        return results